Score Text + QoL Changes
Made the starting platforms have a random position. Players can no longer jump again until they have landed on the top of a platform (rather than touching one at all). Added a temporary material for the platforms so they are easier to see before we get the log sprite in. Added a score text that works off of the height of the camera.